Add an MPI regression for multiple ppn.  Notes:
0) We remove lib directory
1) cartesian comm now inherits from shared memory
2) mpit may be wrong, and is unsupported in this grid version
3) similar to #2, mpi3 is also unsupported
4) the purpose of this commit is to catch up with a year+ of commits
from the main develop branch, but still compile with GRID_COMMS_MPI (not
MPI3).  We want multiple ppn due to contraction tunings (which we have
so far been unable to replicate for ppn = 1).  Additionally, the mpi3
shared mem options for multiple ppn has only map anon, which will not
allow for memory safety when using a shared mem comms buffer (from my
recollection).  We hope this commit might be useful to other projects
which are forced to use multiple ppn (esp. for comms on the KNL.  We
should recall that the KNL comms may eventually be solved using
multi-endpoint, so there may be a limited window for the usefulness of this snapshot.)

More testing is needed.  Caveat Emptor.